+++ src/Image/config/services.php
@@ -19,12 +19,12 @@ use Symfony\UX\Image\ConfigurationValidator;
 use Symfony\UX\Image\DataCollector\ImageConfigurationCollector;
 use Symfony\UX\Image\ProcessingLimits;
 use Symfony\UX\Image\Processor\ChainImageProcessor;
-use Symfony\UX\Image\Processor\GdImageProcessor;
+use Symfony\UX\Image\Processor\Gd\GdImageProcessor;
 use Symfony\UX\Image\Processor\ImageDriverInterface;
 use Symfony\UX\Image\Processor\ImageInspector;
 use Symfony\UX\Image\Processor\ImageInspectorInterface;
 use Symfony\UX\Image\Processor\ImageProcessorInterface;
-use Symfony\UX\Image\Processor\InterventionImageProcessor;
+use Symfony\UX\Image\Processor\Intervention\InterventionImageProcessor;
 use Symfony\UX\Image\Profile\ProfileRegistry;
 use Symfony\UX\Image\Regeneration\RegenerationServiceResolver;
 use Symfony\UX\Image\Renderer\DefaultImageRenderer;
